+++ views/404.handlebars
@@ -1,6 +1,6 @@
-{{> /views/layouts/main}}
+{{> main}}
 
-{{!-- Content --}}
+
 <div class="container">
     <div class="row">
         <div class="col-md-12 text-center">
@@ -13,5 +13,4 @@
     </div>
 </div>
 
-{{!-- Footer partial --}}
-{{> /views/partials/footer}}
+{{> footer}}
