+++ views/bookings/edit.handlebars
@@ -32,6 +32,6 @@
     <button type="submit" class="btn btn-primary">Update Booking</button>
 </form>
 
-<a href="/views/dashboard" class="btn btn-secondary">Cancel</a>
+<a href="/dashboard" class="btn btn-secondary">Cancel</a>
 
 {{> footer}}
