+++ views/businesses.handlebars
@@ -1,0 +1,2 @@
+<h1>Welcome to BookMeLocal</h1>
+<p>This is the Business Page for now.</p>
