+++ views/businesses/hours/show.handlebars
@@ -1,0 +1,15 @@
+<div class="container">
+    <h2 class="heading">{{business.name}}</h2>
+
+    <div class="business-details">
+        <h3>Description:</h3>
+        <p>{{business.description}}</p>
+    </div>
+
+    <div class="business-hours">
+        <h3>Business Hours:</h3>
+        <p>{{business.hours}}</p>
+    </div>
+
+    <a href="/businesses/{{business.id}}/edit" class="btn btn-secondary">Edit Business</a>
+</div>
