+++ views/businesses/index.handlebars
@@ -1,14 +1,6 @@
-{{> header}}
-
-<h1>All Businesses</h1>
+<h1>Businesses</h1>
 <ul>
     {{#each businesses}}
-    <li>
-        <h2><a href="/businesses/{{this.id}}">{{this.name}}</a></h2>
-        <p>{{this.description}}</p>
-        <p>Operating Hours: {{this.hours}}</p>
-    </li>
+    <li><a href="/businesses/{{this.id}}">{{this.name}}</a></li>
     {{/each}}
 </ul>
-
-{{> footer}}
