+++ views/businesses/new.handlebars
@@ -1,0 +1,6 @@
+<h1>Add New Business</h1>
+<form action="/businesses" method="post">
+    <input type="text" name="name" placeholder="Business Name" required>
+    <textarea name="description" placeholder="Description" required></textarea>
+    <button type="submit">Create Business</button>
+</form>
