+++ views/businesses/show.handlebars
@@ -1,4 +1,4 @@
-{{> header}}
+<!--{{> header}}
 
 <h1>{{business.name}}</h1>
 <p>{{business.description}}</p>
@@ -15,6 +15,10 @@
 </div>
 {{/if}}
 
-{{!-- You can add more business details here --}}
 
-{{> footer}}
+
+{{> footer}} -->
+
+<h1>{{business.name}}</h1>
+<p>{{business.description}}</p>
+<!-- Display other business details here -->
