+++ views/layouts/main.handlebars
@@ -1,6 +1,6 @@
-{{> /views/partials/header }}
+{{> header }}
 
-{{!-- The body of your main layout goes here, where you'd include the page-specific content --}}
+
 {{#if messages.success_msg}}
 <div class="alert alert-success">{{messages.success_msg}}</div>
 {{/if}}
@@ -16,4 +16,4 @@
 
 {{body}}
 
-{{> /views/partials/footer }}
+{{> footer }}
