+++ views/login.handlebars
@@ -3,5 +3,5 @@
 {{/if}}
 
 <form action="/auth/login" method="POST">
-    <!-- Login form fields -->
+
 </form>
