+++ views/logout.handlebars
@@ -1,0 +1,2 @@
+<h1>Bye BookMeLocal</h1>
+<p>This is the homepage.</p>
