+++ views/services/show.handlebars
@@ -1,0 +1,4 @@
+<h1>{{service.name}}</h1>
+<p>{{service.description}}</p>
+<p>Price: ${{service.price}}</p>
+<a href="/services">Back to Services List</a>
